requirements.txt
nose
numpy
matplotlib.pyplot
os
sys
atexit